Add files via upload
Task 1: Performing Adaptive Binarization and  Cropping of an input image.

The image binarization is a technique used to convert a grayscale image or a colour image into 
a binary image (black and white image).The main goal is to separate the image's foreground 
from its background, typically to highlight certain features or objects in the image or to 
facilitate further processing such as segmentation or object recognition.
Image Binarization is one of the most fundamental ways to extract useful information from a 
given image. 
This can be achieved with the help of a process called thresholding. 
Binarization is a fundamental preprocessing step in various computer vision and image analysis 
applications, particularly in the areas of document analysis, optical character recognition 
(OCR), and pattern recognition.